+++ saleor_app_framework/app_framework/core/mutations.py
@@ -1,0 +1,27 @@
+CREATE_WEBHOOK = """
+mutation WebhookCreate($input: WebhookCreateInput!) {
+  webhookCreate(input: $input) {
+    webhookErrors {
+      field
+      message
+      code
+    }
+    webhook {
+      id
+    }
+  }
+}
+
+"""
+
+VERIFY_TOKEN = """
+mutation TokenVerify($token: String!) {
+  tokenVerify(token: $token) {
+    isValid
+    user {
+      id
+    }
+  }
+}
+
+"""
