+++ saleor_app_framework/app_framework/core/validators.py
@@ -1,27 +1,18 @@
 from .graphql import get_executor, get_saleor_api_url
-
-VERIFY_TOKEN = """
-mutation TokenVerify($token: String!) {
-  tokenVerify(token: $token) {
-    isValid
-    user {
-      id
-    }
-  }
-}
-
-"""
+from .mutations import VERIFY_TOKEN
 
 
 async def verify_token(saleor_domain: str, token: str) -> bool:
     api_url = get_saleor_api_url(saleor_domain)
     executor = get_executor(host=api_url, auth_token=None)
-    response = await executor(
+    response, errors = await executor(
         VERIFY_TOKEN,
         variables={
             "token": token,
         },
     )
+    if errors:
+        return False
     try:
         return response["data"]["tokenVerify"]["isValid"] is True
     except KeyError:
